- Limit campaign image upload count on Front-End.

+++ resources/js/app.js
@@ -35,21 +35,6 @@ $(document).ready(function () {
         $("#light-slider-wrapper").removeClass('d-none');
     });
 
-    if (typeof Dropzone != 'undefined') {
-        Dropzone.options.fileupload = {
-            accept: function (file, done) {
-                if (file.type != "application/vnd.ms-excel" && file.type != "image/jpeg, image/png, image/jpg") {
-                    done("Error! Files of this type are not accepted");
-                } else {
-                    done();
-                }
-            }
-        }
-        Dropzone.options.fileupload = {
-            acceptedFiles: "image/jpeg, image/png, image/jpg"
-        }
-    }
-
     (function ($, window, undefined) {
         "use strict";
 
@@ -58,6 +43,7 @@ $(document).ready(function () {
             if (typeof Dropzone != 'undefined') {
                 if ($("#fileupload").length) {
                     var dz = new Dropzone("#fileupload", {
+                            maxFiles: 3,
                             addRemoveLinks: true,
                             //Handle existing images
                             init: function () {
@@ -133,30 +119,36 @@ $(document).ready(function () {
 
                             dze_info.find('tfoot td').html('<span class="label label-success">' + status.uploaded + ' uploaded</span> <span class="label label-danger">' + status.errors + ' not uploaded</span>');
 
-                            toastr.error('Your File Uploaded Not Successfully!!', 'Error Alert', {
+                            console.error('Your File Uploaded Not Successfully!!', 'Error Alert', {
                                 timeOut: 5000
                             });
-                        }).on('removedfile', function (file) {
+                        })
+                        .on('removedfile', function (file) {
 
-                        //Remove file from server
-                        $.ajax({
-                            headers: {
-                                'X-CSRF-TOKEN': $('meta[name="csrf-token"]').attr('content')
-                            },
-                            url: dz.options.url,
-                            type: 'DELETE',
-                            data: {
-                                'file_id': file.id,
-                            },
-                            complete: function (response) {
-
-                                if (response.status !== 200 || response.responseJSON.status !== 'OK') {
-                                    dz.emit("addedfile", file);
-                                    dz.options.thumbnail.call(dz, file, file.thumbnail_url);
-                                    dz.emit("complete", file);
-                                }
+                            // Case when file exists on ui and not on server
+                            if(!file.id) {
+                                return;
                             }
-                        });
+
+                            //Remove file from server
+                            $.ajax({
+                                headers: {
+                                    'X-CSRF-TOKEN': $('meta[name="csrf-token"]').attr('content')
+                                },
+                                url: dz.options.url,
+                                type: 'DELETE',
+                                data: {
+                                    'file_id': file.id,
+                                },
+                                complete: function (response) {
+
+                                    if (response.status !== 200 || response.responseJSON.status !== 'OK') {
+                                        dz.emit("addedfile", file);
+                                        dz.options.thumbnail.call(dz, file, file.thumbnail_url);
+                                        dz.emit("complete", file);
+                                    }
+                                }
+                            });
                     });
 
                     $('.dropzone-img').each(function () {
